Authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -54,9 +54,10 @@ class QuestionsController extends Controller
      * @param  \App\Question  $questions
      * @return \Illuminate\Http\Response
      */
-    public function show(Question $questions)
+    public function show(Question $question)
     {
-        //
+        $question->increment('views');
+        return view("questions.show", compact('question'));
     }
 
     /**
@@ -67,7 +68,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies("update-question", $question)) {
+            abort(403, "Access Denied!");
+        } 
         return view("questions.edit", compact('question'));
+        
     }
 
     /**
@@ -92,6 +97,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies("delete-question", $question)) {
+            abort(403, "Access Denied!");
+        } 
         $question->delete();
 
         return redirect()->route("questions.index")->with("success", "Your question has been deleted.");
